MNT: Rename corr_gen to correlation
